requirements.txt
cloudpickle==3.1.1
Farama-Notifications==0.0.4
filelock==3.18.0
fsspec==2025.5.1
gymnasium==1.1.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
PyYAML==6.0.2
sympy==1.14.0
torch==2.7.0
typing_extensions==4.13.2
